WebView Component path refactor
