+++ app/page/Components/OutsideClickHandler/OutsideClickHandler.scss
@@ -1,0 +1,7 @@
+.OutsideClickHandler {
+  height: 100%;
+  width: 100%;
+
+  display: flex;
+  justify-content: flex-end;
+}
