+++ app/page/webview/components/Button/HistoryGoBackButton/HistoryGoBackButton.scss
