+++ app/page/webview/components/WebWidgetMobileHeader/WebWidgetMobileHeader.scss
